+++ src/constant/index.ts
@@ -1,0 +1,1 @@
+export const PAGE_LIMIT = 4;
